# Remove blocksize_distributed split on y