Remove blocking on the Y axis from CPU and GPU schedules, in an attempt to remove constraints on local parallelism.

CPU performance/scalability seems about the same, in my testing.  (See below)

GPU schedule was simplified a bit more, to remove blocking on the Y axis, and remove my misuse of the `gpu_lanes` directive.  Performance is about 40% slower.  There are further improvements possible here, but it's not a priority.

Distribution pattern should be identical between CPU and GPU.

Here are the CPU runtimes I got from testing with the Sandia dataset, 4 threads per node.

```
before patch:
2 nodes:
Halide backprojection returned 0 in 42765 ms
Halide backprojection returned 0 in 43119 ms
4 nodes:
Halide backprojection returned 0 in 21077 ms
Halide backprojection returned 0 in 21176 ms
Halide backprojection returned 0 in 24639 ms
Halide backprojection returned 0 in 24984 ms
8 nodes:
Halide backprojection returned 0 in 11091 ms
Halide backprojection returned 0 in 11414 ms
Halide backprojection returned 0 in 11608 ms
Halide backprojection returned 0 in 12119 ms
Halide backprojection returned 0 in 12187 ms
Halide backprojection returned 0 in 12281 ms
Halide backprojection returned 0 in 14136 ms
Halide backprojection returned 0 in 14310 ms

after patch:
2 nodes:
Halide backprojection returned 0 in 43549 ms
Halide backprojection returned 0 in 45614 ms
4 nodes:
Halide backprojection returned 0 in 20503 ms
Halide backprojection returned 0 in 20947 ms
Halide backprojection returned 0 in 24490 ms
Halide backprojection returned 0 in 25900 ms
8 nodes:
Halide backprojection returned 0 in 11230 ms
Halide backprojection returned 0 in 11355 ms
Halide backprojection returned 0 in 11506 ms
Halide backprojection returned 0 in 11506 ms
Halide backprojection returned 0 in 12184 ms
Halide backprojection returned 0 in 12421 ms
Halide backprojection returned 0 in 14449 ms
Halide backprojection returned 0 in 16613 ms
```