@cimes-isi 2021-02-19
I ran this on Summit for 32kx32k at 256 nodes with the distributed CPU schedule.  There is some improvement in thread utilization, but not as much as I'd hoped.  It was about 31-32 threads for `img` and now is about 32-34.  Total runtime improved from 1526 s to 1417 s.

I might merge this later, but I think we should further discuss/experiment.

Note also: since `blocksize_distributed` is not longer used here, it should be removed entirely.