@cimes-isi 2021-02-19
[SARBP-Sandia-256.746334.log](https://github.com/ISI-apex/halide-sar-app/files/6012080/SARBP-Sandia-256.746334.log)
[SARBP-Sandia-256-no_blocksize_distributed.749331.log](https://github.com/ISI-apex/halide-sar-app/files/6012081/SARBP-Sandia-256-no_blocksize_distributed.749331.log)