@Infinoid 2021-02-19
Looking at the last few nodes in the log, I see that the thread parallelism hovering between 31-38 for all Funcs, whether it's distributed or not.  In fact, the only "threads: 42.0" I see are for really tiny things, <5ms.

Do you observe the same pattern for other apps?