@Infinoid 2021-02-19
Ok.  Here's how it looks to me:

We had a practical limit on parallelism of 32 threads.  The profiles reflect this:

```
1: 167:   norm_rr0:              27424.931ms(5%)   threads: 31.999 peak: 33554432 num: 1048576   avg: 1048576
1: 167:   img:                   445488.750ms(89%) threads: 32.000 stack: 64
1: 224:   norm_rr0:              56829.039ms(11%)  threads: 30.149 peak: 33554432 num: 1048576   avg: 1048576
1: 224:   img:                   425729.906ms(84%) threads: 30.000 stack: 64
1: 174:   norm_rr0:              37198.226ms(7%)   threads: 31.169 peak: 33554432 num: 1048576   avg: 1048576
1: 174:   img:                   422619.937ms(88%) threads: 31.175 stack: 64
1:   0:   norm_rr0:              27218.693ms(5%)   threads: 31.999 peak: 33554432 num: 1048576   avg: 1048576
1:   0:   img:                   446599.906ms(90%) threads: 31.999 stack: 64
```

Close to 32, but never above 32.

This patch lifts that restriction, but it didn't help that much.

```
1:  84:   norm_rr0:              65782.867ms(15%)  threads: 35.729 peak: 44040192 num: 1048576   avg: 1048576
1:  84:   img:                   352502.406ms(80%) threads: 33.031 stack: 64
1: 249:   norm_rr0:              37135.921ms(8%)   threads: 33.327 peak: 44040192 num: 1048576   avg: 1048576
1: 249:   img:                   389488.031ms(86%) threads: 32.499 stack: 64
1: 117:   norm_rr0:              38079.222ms(7%)   threads: 32.351 peak: 44040192 num: 1048576   avg: 1048576
1: 117:   img:                   419290.718ms(87%) threads: 34.343 stack: 64
1:   0:   norm_rr0:              27763.125ms(6%)   threads: 32.092 peak: 44040192 num: 1048576   avg: 1048576
1:   0:   img:                   409558.718ms(89%) threads: 32.293 stack: 64
```

Generally greater than 32, but not by much.

So it seems like something else is going on.  Do you get good thread utilization for single node jobs?